+++ charts/manticoresearch/CHANGELOG.md
@@ -1,3 +1,6 @@
-### 6.2.124
+### 6.2.130
 
-* Added support for priorityClassName
+* Start using domains in replication instead IP 
+* Changed binlog dir (remove data folder from default path)
+* Readiness probe has been removed
+* Added check for max allowed length of worker hostname